refact: extracts comment and data_request_status models to their own files
